move some struct to a new file.
