+++ cdc/kv/metrics.go
@@ -91,7 +91,7 @@ var (
 			Namespace: "ticdc",
 			Subsystem: "kvclient",
 			Name:      "batch_resolved_event_size",
-			Help:      "The number of region in one batch resolved ts event",
+			Help:      "The number of region in one batch resolved checkpointTs event",
 			Buckets:   prometheus.ExponentialBuckets(2, 2, 16),
 		}, []string{"namespace", "changefeed"})
 	grpcPoolStreamGauge = prometheus.NewGaugeVec(
